requirements.txt
amqp==1.4.6
aniso8601==0.82
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
billiard==3.3.0.19
boto==2.38.0
bz2file==0.98
celery==3.1.17
certifi==14.5.14
cffi==1.1.0
click==6.2
cluster==1.1.2
colored==1.2.1
cssselect==0.9.1
decorator==4.0.6
dnspython==1.12.0
ecdsa==0.13
elasticsearch==1.5.0
enum34==1.0.4
Fabric==1.10.1
Flask==0.10.1
Flask-RESTful==0.2.12
flower==0.7.3
futures==3.0.2
gensim==0.12.3
geocoder==1.7.1
geopy==1.11.0
gevent==1.0.1
Google-Search-API==1.1.10
goose-extractor==1.0.22
greenlet==0.4.2
gunicorn==19.1.0
httpretty==0.8.6
idna==2.0
ipaddr==2.1.11
ipaddress==1.0.7
ipwhois==0.10.2
ipython==3.1.0
itsdangerous==0.24
jaro-winkler==1.0.2
jieba==0.32
Jinja2==2.7.3
jsonrpclib==0.1.3
jws==0.1.3
kombu==3.0.24
lxml==3.3.5
MarkupSafe==0.23
meld3==1.0.0
milk==0.6.1
ner==0.1
nltk==3.0.0
numpy==1.10.1
openpyxl==2.1.4
paramiko==1.15.2
pexpect==3.3
pika==0.10.0
Pillow==2.9.0
protobuf==2.5.0
psycopg2==2.6
py==1.4.26
pyasn1==0.1.7
pycparser==2.13
pycrypto==2.6.1
pyfiglet==0.7.4
pymongo==2.7.1
pyOpenSSL==0.15.1
pytest==2.7.0
python-jwt==1.1.1
python-Levenshtein==0.12.0
python-logstash==0.4.5
pyttsx==1.1
pytz==2014.10
PyYAML==3.11
ratelim==0.1.6
redis==2.10.3
requests==2.8.1
retrying==1.3.3
reverse-geocoder==1.3
scikit-learn==0.15.0
scipy==0.16.1
selenium==2.47.1
simplejson==3.8.0
six==1.10.0
smart-open==1.3.0
supervisor==3.1.0
termcolor==1.1.0
text-sentence==0.14
textblob==0.9.0
topia.termextract==1.1.0
tornado==4.0.2
torndsession==1.1.4
Unidecode==0.4.18
urllib3==1.10.4
Werkzeug==0.9.6
wheel==0.24.0
zope.interface==4.1.2
